Split the invite mail component
